requirements.txt
Flask>=2.3.0
google-generativeai>=0.3.0
opencv-python>=4.8.0
Pillow>=10.0.0
requests>=2.31.0
Werkzeug>=2.3.0
gunicorn>=21.0.0
